requirements.txt
flask
tensorflow-cpu
Pillow
numpy